chore: implement FillExtrusionLayer in TypeScript
