+++ docs/FillExtrusionLayer.md
@@ -1,4 +1,4 @@
-<!-- This file was autogenerated from FillExtrusionLayer.js do not modify -->
+<!-- This file was autogenerated from FillExtrusionLayer.tsx do not modify -->
 # <MapboxGL.FillExtrusionLayer />
 FillExtrusionLayer is a style layer that renders one or more 3D extruded polygons on the map.
 
@@ -11,10 +11,10 @@ FillExtrusionLayer is a style layer that renders one or more 3D extruded polygon
 | aboveLayerID | `string` | `none` | `false` | Inserts a layer above aboveLayerID. |
 | belowLayerID | `string` | `none` | `false` | Inserts a layer below belowLayerID |
 | layerIndex | `number` | `none` | `false` | Inserts a layer at a specified index |
-| filter | `array` | `none` | `false` | Filter only the features in the source layer that satisfy a condition that you define |
-| minZoomLevel | `number` | `none` | `false` | The minimum zoom level at which the layer gets parsed and appears. |
-| maxZoomLevel | `number` | `none` | `false` | The maximum zoom level at which the layer gets parsed and appears. |
-| style | `union` | `none` | `false` | Customizable style attributes |
+| filter | `FilterExpression` | `none` | `false` | Filter only the features in the source layer that satisfy a condition that you define |
+| minZoomLevel | `number` | `none` | `true` | The minimum zoom level at which the layer gets parsed and appears. |
+| maxZoomLevel | `number` | `none` | `true` | The maximum zoom level at which the layer gets parsed and appears. |
+| style | `FillExtrusionLayerStyleProps` | `none` | `false` | Customizable style attributes |
 
 
 ## styles
